requirements.txt
tensorflow==2.15.0
tensorflowjs==4.14.0
tf2onnx==1.15.1
onnx==1.15.0
onnxruntime==1.16.3
numpy==1.24.3
pillow==10.1.0
matplotlib==3.8.2
scikit-learn==1.3.2
pandas==2.1.4
kaggle==1.5.16

# Google Drive API
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.111.0
